% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTSPred.R
\docType{methods}
\name{BestTSPred}
\alias{BestTSPred}
\alias{BestTSPred}
\alias{BestTSPred,vector-method}
\alias{BestTSPred,StQList-method}
\title{Seleccion de la mejor prediccion de un conjunto de modelos de series temporales.}
\usage{
BestTSPred(x, BestTSPredParam)

BestTSPred(x, BestTSPredParam)

\S4method{BestTSPred}{vector}(x, BestTSPredParam)

\S4method{BestTSPred}{StQList}(x, BestTSPredParam)
}
\arguments{
\item{x}{\code{vector} u objeto de clase \code{StQList}, con las variables y los valores para los
que obtendremos las predicciones.}

\item{BestTSPredParam}{Objeto de clase \linkS4class{BestTSPredParam} con los parametros de los
distintos modelos de prediccion de series temporales y las variables para las que se quieren
obtener las predicciones.}
}
\value{
\code{data.table} con componentes \code{Pred} y \code{STD} que contienen la mejor
prediccion, para cada variable, de entre las especificadas en el objeto de entrada
\code{BestTSPredParam}, y sus correspondientes desviaciones tipicas estimadas.
}
\description{
\code{BestTSPred} Selecciona la mejor prediccion para una serie de tiempo de entre las
proporcionadas por un conjunto de modelos.
}
\details{
Este metodo toma como argumento de entrada el objeto \code{x} que contiene las variables que
queremos predecir y selecciona la mejor prediccion de entre un conjunto de predicciones
especificadas como input en el objeto de clase \linkS4class{BestTSPredParam}.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L))
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list)
BestTSPred(Example1.TS, BestTSPredParam)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
StQList <- readRDS('../E30183.FF.StQList.rds')
Units <- StQ::getUnits(StQList[['MM102016']])
Units <- Units[sample(1:(dim(Units)[1]), 1000)]
data.table::setkeyv(Units, 'NOrden')
StQ::setUnits(StQList) <- Units
VarNames <- c('CifraNeg_13.___', 'Personal_07.__1._1._')
TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L),
                Arima = list('AutoArimaTSPred', forward = 2L))
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list, VarNames = VarNames)
BestTSPred(StQListExample, BestTSPredParam)
}

}
