% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example1.TS.R
\docType{data}
\name{Example1.TS}
\alias{Example1.TS}
\title{Numeric vector.}
\format{Numeric vector of length 24.}
\usage{
data(Example1.TS)
}
\description{
\code{Example1.TS} is a numeric vector with no NA values.
}
