% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example2.TS.R
\docType{data}
\name{Example2.TS}
\alias{Example2.TS}
\title{Numeric vector.}
\format{Numeric vector of length 24.}
\usage{
data(Example2.TS)
}
\description{
\code{Example1.TS} is a numeric vector with some NA values.
}
