% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.BestTSPredParam.R
\docType{methods}
\name{initialize,BestTSPredParam-method}
\alias{initialize,BestTSPredParam-method}
\title{Inicializador de objetos de clase \linkS4class{BestTSPredParam}.}
\usage{
\S4method{initialize}{BestTSPredParam}(.Object,
  TSPred.list = stop("[BestTSPredParam initialize] Es necesario especificar una lista de modelos de predicci蚤."),
  VarNames = NULL)
}
\arguments{
\item{.Object}{Par치metro obligatorio para todos los inicializadores.}

\item{TSPred.list}{Objeto de clase \code{list} cuyas componentes contienen los
nombres y los par치metros de los modelos con los que se generan las
distintas predicciones de los datos.}

\item{VarNames}{\code{Vector} de tipo \code{character} con los nombres de las
variables a predecir. Por defecto, toma el valor \code{NULL}}
}
\value{
Objeto de clase \linkS4class{BestTSPredParam}.
}
\description{
Crea un objeto de clase \linkS4class{BestTSPredParam} con los par치metros que
utiliza como argumentos el m칠todo \linkS4class{BestTSPred}.
}
\details{
Esta funci칩n construye un objeto de clase \code{BestTSPredParam} a partir
de una lista que contiene los nombres y los par치metros de los distintos
modelos de predicci칩n de series temporales con los que se obtienen predicciones,
los nombres de las variables a predecir y los nombres de las variables que se
utilizar치n como variables clave.
}
\examples{
TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L))
BestTSPredParam1 <- new(Class='BestTSPredParam',
                       TSPred.list = TS.list)
str(BestTSPredParam1)

BestTSPredParam2 <- new(Class='BestTSPredParam',
                       TSPred.list = TS.list[1:2])
str(BestTSPredParam2)

}

