% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTSPred.R
\docType{methods}
\name{BestTSPred}
\alias{BestTSPred}
\alias{BestTSPred,StQList-method}
\alias{BestTSPred,vector-method}
\title{Selección de la mejor predicción de un conjunto de modelos de series temporales.}
\usage{
BestTSPred(x, BestTSPredParam)

\S4method{BestTSPred}{vector}(x, BestTSPredParam)

\S4method{BestTSPred}{StQList}(x, BestTSPredParam)
}
\arguments{
\item{x}{\code{vector}, \code{matrix} u objeto de clase \code{StQList}, con las variables y los
valores para los que obtenemos las predicciones.}

\item{BestTSPredParam}{Objeto de clase \linkS4class{BestTSPredParam} con los parámetros
de los distintos modelos de predicción de series temporales con los que se generan
las diferentes predicciones y las variables sobre las que se calculan.}
}
\value{
Lista con componentes \code{Pred} y \code{STD} que contienen
la mejor predicción de entre las especificadas en el objeto de entrada
\code{BestTSPredParam}, y sus correspondientes desviaciones típicas estimadas,
respectivamente.

\itemize{
 \item For input class vector, it returns numeric vectors.
 \item For input class matrix, it returns matrices.
 \item For input class StQList, it returns list whose components are
  data.tables.
}
}
\description{
\code{BestTSPred} Selecciona la mejor predicción para una serie de tiempo de
entre las proporcionadas por un conjunto de modelos.
}
\details{
Este método toma como argumento de entrada el objeto \code{x} que contiene las
variables que queremos predecir y selecciona la mejor predicción de entre un
conjunto de predicciones especificadas como input en el objeto de clase
\linkS4class{BestTSPredParam}. Este parámetro contiene los diferentes modelos
de predicción de series temporales que generan las distintas predicciones.
}
\examples{

# Predicting one and two months ahead in time
data(Example1.TS)
TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L))
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list)
BestTSPred(Example1.TS, BestTSPredParam)

\dontrun{
# With an object of class StQList
data(StQListExample)
VarNames <- c('ActivEcono_35._6._2.1.4._0', 'GeoLoc_35._6._2.1._1.2.5.')
TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L),
                Arima = list('AutoArimaTSPred', forward = 2L))
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list, VarNames = VarNames)
BestTSPred(StQListExample, BestTSPredParam)
}

}

