% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTSPredParam-class.R
\docType{class}
\name{BestTSPredParam-class}
\alias{BestTSPredParam-class}
\title{Clase S4 BestTSPredParam para los parametros del metodo BestTSPred}
\description{
Definicion de la clase S4 \linkS4class{BestTSPredParam} que contiene los parametros que utiliza
el metodo \code{BestTSPred}.
}
\section{Slots}{

\describe{
\item{\code{TSPred.list}}{Objeto de clase \code{\link{list}} cuyos componentes son listas que contienen
el nombre y los argumentos de las funciones con las que se calcularan las predicciones de acuerdo
a los distintos modelos de series temporales.}

\item{\code{VarNames}}{\code{Vector} de tipo \code{character} con los nombres de las variables para las
que se calcularan las predicciones.}
}}

\examples{
# Un prototipo vacío
new(Class = 'BestTSPredParam')

TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L),
                Arima = list('AutoArimaTSPred', forward = 2L))
VarNames <- c('CifraNeg_13.___', 'Personal_07.__1._1._')
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list, VarNames = VarNames)


}
