% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTSPredParam-class.R
\docType{class}
\name{BestTSPredParam-class}
\alias{BestTSPredParam-class}
\title{Clase S4 BestTSPredParam para los parámetros del método BestTSPred}
\description{
Definición de la clase S4 \linkS4class{BestTSPredParam} para los parámetros que
hay que pasar como argumentos a la función \code{BestTSPred}.
}
\details{
La estructura de la clase S4 \code{BestTSPredParam} consta de:
\itemize{
\item el slot \code{TSPred.list} de clase \code{\link{list}} con los argumentos
necesarios para llamar a las funciones de los distintos modelos de series
temporales que generan las predicciones. Se trata de listas cuya primera componente
es el nombre de la función y el resto son los argumentos que requieren dichas
funciones. El output de cada unos de estos modelos debe cumplir con los requisitos
de diseño de las funciones en el paquete \code{TSPred};
\item un slot opcional de tipo \code{character} con los nombres de las variables
a predecir;
}
}
\section{Slots}{

\describe{
\item{\code{TSPred.list}}{Objeto de clase \code{\link{list}} cuyos componentes contienen
los elementos de cada llamada a las funciones que calculan las predicciones de
acuerdo a los distintos modelos.}

\item{\code{VarNames}}{\code{Vector} de tipo \code{character} con los nombres de las
variables a las que corresponden las predicciones.}
}}

\examples{
# Un prototipo vacío
new(Class = 'BestTSPredParam')

TS.list <- list(Reg = list('RegDiffTSPred', forward = 2L),
                Stat = list('StatDiffTSPred', forward = 2L),
                StatReg = list('StatRegDiffTSPred', forward = 2L),
                Arima = list('AutoArimaTSPred', forward = 2L))
VarNames <- c('CifraNeg_13.___', 'Personal_07.__1._1._')
BestTSPredParam <- new(Class='BestTSPredParam', TSPred.list = TS.list, VarNames = VarNames)


}
