% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTSPredParam-class.R
\docType{class}
\name{BestTSPredParam-class}
\alias{BestTSPredParam-class}
\title{Clase S4 BestTSPredParam para los parámetros del método BestTSPred}
\description{
Definición de la clase S4 \linkS4class{BestTSPredParam} para los parámetros que
hay que pasar como argumentos a la función \code{BestTSPred}.
}
\section{Slots}{

\describe{
\item{\code{TSPred.list}}{Objeto de clase \code{\link{list}} cuyos componentes contienen
los elementos de cada llamada a las funciones que calculan las predicciones de
acuerdo a los distintos modelos.}

\item{\code{VarNames}}{\code{Vector} de tipo \code{character} con los nombres de las
variables a las que corresponden las predicciones.}
}}
\examples{
# Un prototipo vacío
new(Class = 'BestTSPredParam')

}

